apps/api/pyproject.toml
[project]
name = "system-design-copilot-api"
version = "0.1.0"
description = "FastAPI service for System Design Co-Pilot"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "pydantic-settings>=2.6,<3",
    "psycopg[binary]>=3.2,<4",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
